services/auth_service/requirements.txt
Django>=4.2
psycopg2-binary>=2.9
djangorestframework
drf-spectacular
google-auth
djangorestframework-simplejwt
requests
pika==1.3.2
pytz>=2023.3